creates Routes file
